Controladores terminados

+++ src/main/java/com/api/pnat/interfaces/IAvistaPorInmersion.java
@@ -3,6 +3,6 @@ package com.api.pnat.interfaces;
 import com.api.pnat.model.AvistaPorInmersion;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IAvistaPorInmersion extends CrudRepository<AvistaPorInmersion, Long> {
+public interface IAvistaPorInmersion extends CrudRepository<AvistaPorInmersion, Integer> {
 
 }
