+++ src/main/java/com/api/pnat/interfaces/ILugares.java
@@ -3,6 +3,6 @@ package com.api.pnat.interfaces;
 import org.springframework.data.repository.CrudRepository;
 import com.api.pnat.model.Lugares;
 
-public interface ILugares extends CrudRepository<Lugares, Long>{
+public interface ILugares extends CrudRepository<Lugares, Integer>{
 
 }
