+++ src/main/java/com/api/pnat/interfaces/IUsuario.java
@@ -3,6 +3,6 @@ package com.api.pnat.interfaces;
 import com.api.pnat.model.Usuario;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IUsuario extends CrudRepository<Usuario, Long>{
+public interface IUsuario extends CrudRepository<Usuario, Integer>{
 
 }
