+++ src/main/java/com/api/pnat/interfaces/IUsuarioPorInmersion.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.api.pnat.model.UsuarioPorInmersion;
 
-public interface IUsuarioPorInmersion extends CrudRepository<UsuarioPorInmersion, Long>{
+public interface IUsuarioPorInmersion extends CrudRepository<UsuarioPorInmersion, Integer>{
 
 }
